requirements.txt
streamlit>=1.33.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.1
pandas>=2.2.0
